// src/services/github/openapi.d.ts
export interface components {
    schemas: {
// ...
        "basic-error": {
            message?: string;
            documentation_url?: string;
            url?: string;
            status?: string;
        };
// ...
        issue: {
            /** Format: int64 */
            id: number;
            node_id: string;
            /**
             * Format: uri
             * @description URL for the issue
             * @example https://api.github.com/repositories/42/issues/1
             */
            url: string;
            /** Format: uri */
            repository_url: string;
            labels_url: string;
            /** Format: uri */
            comments_url: string;
            /** Format: uri */
            events_url: string;
            /** Format: uri */
            html_url: string;
            /**
             * @description Number uniquely identifying the issue within its repository
             * @example 42
             */
            number: number;
            /**
             * @description State of the issue; either 'open' or 'closed'
             * @example open
             */
            state: string;
            /**
             * @description The reason for the current state
             * @example not_planned
             * @enum {string|null}
             */
            state_reason?: "completed" | "reopened" | "not_planned" | "duplicate" | null;
            /**
             * @description Title of the issue
             * @example Widget creation fails in Safari on OS X 10.8
             */
            title: string;
            /**
             * @description Contents of the issue
             * @example It looks like the new widget form is broken on Safari. When I try and create the widget, Safari crashes. This is reproducible on 10.8, but not 10.9. Maybe a browser bug?
             */
            body?: string | null;
            user: components["schemas"]["nullable-simple-user"];
            /**
             * @description Labels to associate with this issue; pass one or more label names to replace the set of labels on this issue; send an empty array to clear all labels from the issue; note that the labels are silently dropped for users without push access to the repository
             * @example [
             *       "bug",
             *       "registration"
             *     ]
             */
            labels: (string | {
                /** Format: int64 */
                id?: number;
                node_id?: string;
                /** Format: uri */
                url?: string;
                name?: string;
                description?: string | null;
                color?: string | null;
                default?: boolean;
                /** @description The user who archived the label, or `null` if it has not been archived. */
                archived_by?: components["schemas"]["simple-user"] | null;
            })[];
            assignees?: components["schemas"]["simple-user"][];
            milestone: components["schemas"]["nullable-milestone"];
            locked: boolean;
            active_lock_reason?: string | null;
            comments: number;
            pull_request?: {
                /** Format: date-time */
                merged_at?: string;
                /** Format: uri */
                diff_url: string;
                /** Format: uri */
                html_url: string;
                /** Format: uri */
                patch_url: string;
                /** Format: uri */
                url: string;
            };
            /** Format: date-time */
            closed_at: string | null;
            /** Format: date-time */
            created_at: string;
            /** Format: date-time */
            updated_at: string;
            draft?: boolean;
            closed_by?: components["schemas"]["nullable-simple-user"];
            body_html?: string;
            body_text?: string;
            /** Format: uri */
            timeline_url?: string;
            type?: components["schemas"]["issue-type"];
            repository?: components["schemas"]["repository"];
            performed_via_github_app?: components["schemas"]["nullable-integration"];
            author_association?: components["schemas"]["author-association"];
            reactions?: components["schemas"]["reaction-rollup"];
            sub_issues_summary?: components["schemas"]["sub-issues-summary"];
            /**
             * Format: uri
             * @description URL to get the parent issue of this issue, if it is a sub-issue
             */
            parent_issue_url?: string | null;
            pinned_comment?: components["schemas"]["nullable-issue-comment"];
            issue_dependencies_summary?: components["schemas"]["issue-dependencies-summary"];
            issue_field_values?: components["schemas"]["issue-field-value"][];
        };
        /**
         * Issue Comment
         * @description Comments provide a way for people to collaborate on an issue.
         */
        "issue-comment": {
            /**
             * Format: int64
             * @description Unique identifier of the issue comment
             * @example 42
             */
            id: number;
            node_id: string;
            /**
             * Format: uri
             * @description URL for the issue comment
             * @example https://api.github.com/repositories/42/issues/comments/1
             */
            url: string;
            /**
             * @description Contents of the issue comment
             * @example What version of Safari were you using when you observed this bug?
             */
            body?: string;
            body_text?: string;
            body_html?: string;
            /** Format: uri */
            html_url: string;
            user: components["schemas"]["nullable-simple-user"];
            /**
             * Format: date-time
             * @example 2011-04-14T16:00:49Z
             */
            created_at: string;
            /**
             * Format: date-time
             * @example 2011-04-14T16:00:49Z
             */
            updated_at: string;
            /** Format: uri */
            issue_url: string;
            author_association?: components["schemas"]["author-association"];
            performed_via_github_app?: components["schemas"]["nullable-integration"];
            reactions?: components["schemas"]["reaction-rollup"];
            pin?: components["schemas"]["nullable-pinned-issue-comment"];
            minimized?: components["schemas"]["nullable-issue-comment-minimized"];
        };
// ...
    };
// ...
}